+++ featurehub-sdk/CHANGELOG.md
@@ -1,6 +1,16 @@
+## [1.2.0] - 2022-10-14
+
+- Fixing polling client for server eval
+- Added in tests for polling client
+- Added in server eval for streaming client
+- Added expired environment support for polling and streaming
+- Added cache busting for server eval polling client
+- 
+
 ## [1.1.0] - 2022-10-12
 
 - Adds support for array values in flag evaluation contexts via [this PR](https://github.com/featurehub-io/featurehub-ruby-sdk/pull/12)
+
 ## [1.0.0] - 2022-06-06
 
 - Initial release, feature complete